Analise_Mercado/requirements.txt
crewai  
crewai_tools
langchain_community
ipykernel
python-dotenv
jupyter
pdfkit
markdown